third_party/blink/web_tests/external/wpt/html/browsers/the-window-object/window-open-noopener_indexed-expected.txt

This is a testharness.js-based test.
FAIL window.open() with 'noopener' should reuse existing target assert_equals: Should still have opener expected true but got false
FAIL noopener=1 means the same as noopener assert_equals: Should still have opener expected true but got false
PASS noopener=0 means lack of noopener
FAIL noopener separated only by spaces should work assert_equals: Should still have opener expected true but got false
FAIL Trailing noopener should work assert_equals: Should still have opener expected true but got false
FAIL Leading noopener should work assert_equals: Should still have opener expected true but got false
FAIL Interior noopener should work assert_equals: Should still have opener expected true but got false
FAIL noreferrer should also suppress opener when reusing existing target assert_equals: Should still have opener expected true but got false
Harness: the test ran to completion.
